small change to main and move example files
