+++ packages/webview_flutter/webview_flutter_wkwebview/example/integration_test/legacy/webview_flutter_test.dart
@@ -21,9 +21,9 @@ import 'package:integration_test/integration_test.dart';
 import 'package:webview_flutter_platform_interface/webview_flutter_platform_interface_legacy.dart';
 import 'package:webview_flutter_wkwebview/src/common/instance_manager.dart';
 import 'package:webview_flutter_wkwebview/src/common/weak_reference_utils.dart';
-import 'package:webview_flutter_wkwebview_example/navigation_decision.dart';
-import 'package:webview_flutter_wkwebview_example/navigation_request.dart';
-import 'package:webview_flutter_wkwebview_example/web_view.dart';
+import 'package:webview_flutter_wkwebview_example/legacy/navigation_decision.dart';
+import 'package:webview_flutter_wkwebview_example/legacy/navigation_request.dart';
+import 'package:webview_flutter_wkwebview_example/legacy/web_view.dart';
 
 Future<void> main() async {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
